% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nooa.R
\name{convert_data}
\alias{convert_data}
\title{convert_data}
\usage{
convert_data(Year = 0, Month = 0, Day = 0)
}
\arguments{
\item{Year}{An integer vector representing years}

\item{Month}{An integer vector representing months}

\item{Day}{An integer vector representing days}
}
\value{
A date vector or an error message if the input parameters
don't have the same length
}
\description{
A date vector is created by uniting the year, month, day and
converting it to the Date class
}
\examples{
Year <- c(-11, -100, -200, -300, 2000, 0)
Month <- c(NA, NA, NA, 9, 8, 1)
Day <- c(8, NA, 7, 29, 20, 12)

convert_data(Year, Month, Day)

}
