% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nooa.R
\name{coords2country}
\alias{coords2country}
\title{coords2country}
\usage{
coords2country(latitude, longitude)
}
\arguments{
\item{latitude}{A numeric value representing latitude}

\item{longitude}{A numeric value representing longitude}
}
\value{
a string representing the name of the Country. The NA is returned
if the coordinates (logitude and latitude) is not in a country.
}
\description{
Finds the country where the location is located
by \code{latitude} and \code{longitude}.
}
\details{
I made small modification to the function \strong{coords2continent}. See the location:
\url{https://stackoverflow.com/questions/21708488/get-country-and-continent-from-longitude-and-latitude-point-in-r}
Thanks to the user \strong{Andy} for the function \strong{coords2continent}.
}
\examples{
country <- coords2country(33.40,73.30)

}
