% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nooa.R
\name{create_map}
\alias{create_map}
\title{create_map}
\usage{
create_map(df, content)
}
\arguments{
\item{df}{A NOAA dataframe}

\item{content}{a string vector which contains HTML code}
}
\value{
A leaflet object (a map)
}
\description{
The function draw the epicenters eartquakes
(LATITUDE/LONGITUDE) from a filtered NOAA dataframe.
Each point is annotates with an pop up window containing
annotation data stored in a column of the data frame.
}
\details{
Using the columns from filtered frameframe,
we create (for each location on the map) an HtML string
(which is displayed in the pop-up window).
}
\examples{
df <- load_data()
df <- eq_clean_data(df)
df <- eq_location_clean(df)

dff <- dplyr::filter(df, COUNTRY=="CHINA" & YEAR>=2010)
magnitude <- ifelse(is.na(dff$EQ_PRIMARY),4,dff$EQ_PRIMARY)

content <- paste("<b>Location:</b>", df$LOCATION_NAME,"<br/>",
                 "<b>Magnitude:</b>", df$EQ_PRIMARY,"<br/>",
                 "<b>Total deaths:</b>", df$TOTAL_DEATHS)

a <- create_map(dff, content)

}
