% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nooa.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{A dataframe containing NOAA Earthquake data}
}
\value{
A cleaned date frame
}
\description{
A new dateframe is created (a cleaned dataframe) from a raw dataframe
by uniting the year, month, day in a single Date column (named DataEQ)
and converting some columns (LATITUDE, LONGITUDE, EQ_PRIMARY and DEATHS)
to numeric.Also I realized data imputation
using `fix_na...` functions.
}
\examples{
df <- load_data()
df <- eq_clean_data(df)

}
