% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nooa.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{A dataframe containing NOAA Earthquake data}
}
\value{
A dataframe with column LOCATION_NAME modified
}
\description{
Cleans the LOCATION_NAME column by stripping out
the country name (including the colon) and converts names to
title case (as opposed to all caps).
}
\examples{
dfnew <- load_data()
dff <- eq_location_clean(dfnew)

}
