% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nooa.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(df, annot_col = "popup text")
}
\arguments{
\item{df}{A NOAA dataframe filterd after COUNTRY and YEAR}

\item{annot_col}{A string representing the name of column
used to annotate earthquake marker}
}
\value{
A a map displaying earthquate location for a given country with user defined popup
}
\description{
The function show on a map eartquakes for a
particular country
}
\examples{

df <- load_data()
df <- eq_clean_data(df)
df <- dplyr::filter(df, COUNTRY == "MEXICO" & lubridate::year(DataEQ) >= 2000)
g <- eq_map(df, annot_col = "DataEQ")

}
