% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nooa.R
\name{filterData}
\alias{filterData}
\title{filterData}
\usage{
filterData(df, COUNTRY, minYear = 2000, maxYear = 2001)
}
\arguments{
\item{df}{a NOOA Dataframe}

\item{COUNTRY}{a vector of characters representing
country after which we will filter}

\item{minYear}{a minimum value for YEAR}

\item{maxYear}{a maximum value for YEAR}
}
\value{
a filtered dataframe
}
\description{
A cleaned NOOA dataframe is prepared for
visualisation. That means that the eartquakes
are filtered after a vector of \code{COUNTRY}  and
a period of time. The period of time is given between
a \code{minYear} and \code{maxYear}.
}
\examples{
df <- load_data()
df <- eq_clean_data(df)
df <- eq_location_clean(df)
dff <- filterData(df, COUNTRY="CUBA", minYear=1880, maxYear=1992)

}
