% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nooa.R
\name{fix_countries}
\alias{fix_countries}
\title{fix_countries}
\usage{
fix_countries(dff)
}
\arguments{
\item{dff}{A dataframe containing NOAA Earthquake data}
}
\value{
A date frame with the values from COUNTRY corrected
}
\description{
Checks that the country in which the location
(given by the columns LATITUDE and LONGITUDE) is the same with the
value from COUNTRY column. If there are differences, the value
from the COUNTRY column is updated.
}
\details{
To determine the location's country,
we use the coords2country function. If the function coords2country
return NA, we don't update the value from the COUNTRY column.
}
\examples{
dfnew <- load_data()
dfnew <- dplyr::mutate(dfnew, LATITUDE=as.numeric(LATITUDE))
dfnew <- dplyr::mutate(dfnew, LONGITUDE=as.numeric(LONGITUDE))
dfnew <- dplyr::mutate(dfnew, EQ_PRIMARY=as.numeric(EQ_PRIMARY))
dfnew <- dplyr::mutate(dfnew, DEATHS=as.numeric(DEATHS))
dfnew <- fix_na_latitude(dfnew)
dfnew <- fix_na_longitude(dfnew)
dfnew <- fix_na_lon_na_lat(dfnew)
dfnew <- dplyr::mutate(dfnew, DataEQ=convert_data(YEAR, MONTH, DAY))
test <- fix_countries(dfnew)

}
