% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nooa.R
\name{fix_na_latitude}
\alias{fix_na_latitude}
\title{fix_na_latitude}
\usage{
fix_na_latitude(dff)
}
\arguments{
\item{dff}{A dataframe containing NOAA Earthquake data}
}
\value{
A date frame (with those cases corrected)
}
\description{
Correct the cases with LATITUDE is NA and LONGITUDE is not NA.
}
\details{
Information about those earthquakes was found in the next documents (sites):
\itemize{
\item \url{https://www.researchgate.net/publication/296476641_A_catalogue_of_earthquakes_between_810BC_and_2012_for_the_Bay_of_Bengal}
\item \url{http://www.unesco-uganda.ug/files/downloads/Ethquakes\%20in\%20Uganda.pdf} (see pg.3)
\item \url{https://en.wikipedia.org/wiki/List_of_earthquakes_in_1939}
}
}
\examples{

dfnew <- load_data()
dfnew <- dplyr::mutate(dfnew, LATITUDE=as.numeric(LATITUDE))
dfnew <- dplyr::mutate(dfnew, LONGITUDE=as.numeric(LONGITUDE))
dfnew <- dplyr::mutate(dfnew, EQ_PRIMARY=as.numeric(EQ_PRIMARY))
dfnew <- dplyr::mutate(dfnew, DEATHS=as.numeric(DEATHS))
dfnew <- fix_na_latitude(dfnew)

}
