% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nooa.R
\name{fix_na_lon_na_lat}
\alias{fix_na_lon_na_lat}
\title{fix_na_lon_na_lat}
\usage{
fix_na_lon_na_lat(dff)
}
\arguments{
\item{dff}{A dataframe containing NOAA Earthquake data}
}
\value{
A date frame (with those cases corrected)
}
\description{
Correct the cases with LATITUDE is NA and LONGITUDE is NA.
}
\details{
Information about the location of those earthquakes was found using the
geocode function from ggmap package. The dataframe generated by
the geocode function was saved in the file "nulllatlong.rds".
For the cases where the geocode function could not find the location of the earthquakes,
the following documents were used (sites):
\itemize{
\item \url{ftp://ftp.ngdc.noaa.gov/hazards/publications/Wdcse-49.pdf} (pg. 248)
\item \url{https://mapcarta.com/32906862}
\item \url{http://www.jonfr.com/volcano/?p=745}
\item \url{https://mapcarta.com/19182696}
\item \url{http://www.pmd.gov.pk/Islamabadreport.pdf} (pg 22)
\item \url{https://arxiv.org/ftp/arxiv/papers/1708/1708.07262.pdf} (pg. 4)
\item \url{https://celt.ucc.ie/published/T100010A/text006.html}
}
}
\examples{
dfnew <- load_data()
dfnew <- dplyr::mutate(dfnew, LATITUDE=as.numeric(LATITUDE))
dfnew <- dplyr::mutate(dfnew, LONGITUDE=as.numeric(LONGITUDE))
dfnew <- dplyr::mutate(dfnew, EQ_PRIMARY=as.numeric(EQ_PRIMARY))
dfnew <- dplyr::mutate(dfnew, DEATHS=as.numeric(DEATHS))
dfnew <- fix_na_lon_na_lat(dfnew)

}
