% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nooa.R
\name{fix_na_longitude}
\alias{fix_na_longitude}
\title{fix_na_longitude}
\usage{
fix_na_longitude(dff)
}
\arguments{
\item{dff}{A dataframe containing NOAA Earthquake data}
}
\value{
A date frame (with those cases corrected)
}
\description{
Correct the cases with LATITUDE is not NA and LONGITUDE is NA.
}
\details{
Information about those earthquakes was found in the next documents (sites):
\itemize{
\item \url{https://en.wikipedia.org/wiki/List_of_earthquakes_in_the_British_Isles}
}
}
\examples{
dfnew <- load_data()
dfnew <- dplyr::mutate(dfnew, LATITUDE=as.numeric(LATITUDE))
dfnew <- dplyr::mutate(dfnew, LONGITUDE=as.numeric(LONGITUDE))
dfnew <- dplyr::mutate(dfnew, EQ_PRIMARY=as.numeric(EQ_PRIMARY))
dfnew <- dplyr::mutate(dfnew, DEATHS=as.numeric(DEATHS))
dfnew <- fix_na_longitude(dfnew)

}
