% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nooa.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}

\item{...}{...}
}
\value{
ggplot2 graphical object
}
\description{
This geom adds a vertical line to
each data point with a text annotation
(e.g. the location of the earthquake)
attached to each line. and number of deaths.
}
\examples{
df <- load_data()
df <- eq_clean_data(df)
df <- eq_location_clean(df)

country <- c("JAPAN")
dfFilter <- filterData(df, COUNTRY=country, minYear=1000, maxYear=1400)

test <- (ggplot2::ggplot()+
           geom_timeline(data=dfFilter,ggplot2::aes(x=DataEQ,
                                           xmin=as.Date("1000-01-01"),
                                           xmax=as.Date("1400-12-31"),
                                           colour=DEATHS,
                                           fill=DEATHS,
                                           size=EQ_PRIMARY,
                                          y=COUNTRY))+
           geom_timeline_label(data=dfFilter, ggplot2::aes(x=DataEQ,
                                                  location=LOCATION_NAME,
                                                  size=EQ_PRIMARY,
                                                  n_max=5,
                                                  y=COUNTRY)))

}
