% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nooa.R
\docType{data}
\name{geomTimeline}
\alias{geomTimeline}
\title{geomTimeline}
\format{An object of class \code{geomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 6.}
\usage{
geomTimeline
}
\description{
Function to construct a new class for the geom_timeline
}
\section{Aesthetics}{

\code{geom_timeline} understands the following aesthetics:
\itemize{
  \item \code{x} date
  \item \code{y} COUNTRY
  \item \code{xmin} minimum date for earthquakes
  \item \code{xmax} maximum date for earthquakes
  \item \code{size} size of shape has the value of EQ_PRIMARY
  \item \code{fill} fill colour shape (number of deaths) eg DEATHS
  \item \code{colour} colour shape (number of deaths) eg DEATHS
}
}

\keyword{datasets}
