% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nooa.R
\docType{data}
\name{geomTimelineLabel}
\alias{geomTimelineLabel}
\title{geomTimelineLabel}
\format{An object of class \code{geomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}) of length 6.}
\usage{
geomTimelineLabel
}
\description{
geomTimelineLabel
}
\section{Aesthetics}{

\code{geom_timeline} understands the following aesthetics:
\itemize{
  \item \code{x} date
  \item \code{location} name of the location of earthquake
  \item \code{size} size of shape has the value of EQ_PRIMARY
  \item \code{y} country where was the earthquake
  \item \code{n_max} maximux number of lation which is ploted
}
}

\keyword{datasets}
