% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nooa.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data()
}
\value{
A date frame
}
\description{
Load in a dataframe Earthquate data obtained from
https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1
}
\examples{
df <- load_data()

}
