% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nooa.R
\name{remove_countries}
\alias{remove_countries}
\title{remove_countries}
\usage{
remove_countries(df)
}
\arguments{
\item{df}{A dataframe containing NOAA Earthquake data}
}
\value{
A dataframe with column LOCATION_NAME modified
}
\description{
Removes from column LOCATION_NAME (a NOAA dataframe) the strings
which representing a name of country.
}
\examples{
dfnew <- load_data()
dff <- remove_countries(dfnew)

}
