% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nooa.R
\name{setupGeom}
\alias{setupGeom}
\title{setupGeom}
\usage{
setupGeom(p)
}
\arguments{
\item{p}{a ggplot object previously built}
}
\value{
a ggplot object
}
\description{
Create a ggplot theme and add this
to an ggplot object.
}
\examples{
df <- load_data()
df <- eq_clean_data(df)
df <- eq_location_clean(df)
country <- c("JAPAN")
dfFilter <- filterData(df, COUNTRY=country,
minYear=1000, maxYear=1400)
test <- (ggplot2::ggplot()+
           geom_timeline(data=dfFilter,ggplot2::aes(x=DataEQ,
                                           xmin=as.Date("1000-01-01"),
                                           xmax=as.Date("1400-12-31"),
                                           colour=DEATHS,
                                           fill=DEATHS,
                                           size=EQ_PRIMARY,
                                           y=COUNTRY))+
           geom_timeline_label(data=dfFilter, ggplot2::aes(x=DataEQ,
                                                  location=LOCATION_NAME,
                                                  size=EQ_PRIMARY,
                                                  n_max=5,
                                                  y=COUNTRY)))

a <- setupGeom(test)

}
